chore: nav-home을 사용하는 레이아웃 그룹핑
