+++ app/(main)/explore/page.tsx
@@ -1,0 +1,9 @@
+const Explore = () => {
+  return (
+    <div>
+      <h1>explore</h1>
+    </div>
+  );
+};
+
+export default Explore;
