+++ app/(main)/layout.tsx
@@ -1,0 +1,16 @@
+import type { ReactNode } from "react";
+
+import NavHome from "@/shared/ui/nav-home/nav-home";
+
+export default function MainLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <div>
+      <NavHome />
+      {children}
+    </div>
+  );
+}
