+++ app/(main)/my/page.tsx
@@ -1,0 +1,9 @@
+const My = () => {
+  return (
+    <div>
+      <h1>my</h1>
+    </div>
+  );
+};
+
+export default My;
