+++ app/(main)/setting/page.tsx
@@ -1,0 +1,9 @@
+const Setting = () => {
+  return (
+    <div>
+      <h1>setting</h1>
+    </div>
+  );
+};
+
+export default Setting;
