--- app/my/page.tsx
@@ -1,12 +1,0 @@
-import NavHome from "@/shared/ui/nav-home/nav-home";
-
-const My = () => {
-  return (
-    <div style={{ height: "100%", width: "100%" }}>
-      <NavHome />
-      <h1 style={{ marginTop: "6.4rem" }}>my</h1>
-    </div>
-  );
-};
-
-export default My;
